+++ src/interface/api/api_routes.h
@@ -27,5 +27,7 @@ void registerWifiRoutes(AsyncWebServer &server);
 void registerCalibrationRoutes(AsyncWebServer &server);
 void registerHistoryRoutes(AsyncWebServer &server);
 void registerPumpRoutes(AsyncWebServer &server);
+void registerProcessRoutes(AsyncWebServer &server);
 void registerProfilesRoutes(AsyncWebServer &server);
+void registerSafetyRoutes(AsyncWebServer &server);
 void registerTestingRoutes(AsyncWebServer &server);